Feat/merge opeps datasets (#120)
* added general priors

* added ocapeptides dataset (WIP)

* proirs WIP

* multidataset tools

* cleanup and import fix

* added split saving

* arr fix

* added splitter

* tqdm desc fix

* refactor

* fixed CATH get function call

* added chunker

* type hints/doc

* typing imports

* YAML and chunker update

* target generation and CATH CACB swap

* print removal

* delta force shape fix

* boolean/nonetype fix

* training tools

* add simulation examples

* gradient accumulation

* gradient accumulation fix

* merge fix

* dihedral ajustments

* dihedral multi data tools wip

* generator routines update

* cleanup and wip

* dihedral cleanup WIP

* dihedral constrain fix

* alternate dihedral tag fix

* additional compute names

* adjusted Dihedral attributes and forward

* generator wip

* dihedral -> torsions

* Dihedral forward fix

* dihedral buffer fix

* dihedral fix WIP

* adjusted OPEP acc stride

* cath fit param fix

* test_internals_merge

* removed aic test for now

* modifed partition gen

* added mlcg simulation params/script

* separate dihedral from non-dihedral

* temp fix

* wip

* TRPcage target

* add trpcage, amber cln, trpcage sim

* modify dihedral computation

* format

* benchmark sim

* wip

* update doc

* fix pickling issue

* doc

* add the prior specialization to the simulation
classes

* doc

* format

* fix save energies

* import fix

* file reorganization

* added embeddings file

* try improving handling of exploding trajectories

* added trpcage loader

* fixed get_trpcage function

* VERY IMPORTANT: fixed omega fitting range

* added ShiftedPeriodicHarmonicImpropers

* adapted transferable baseline fit and format

* fixed prior assignment

* adjusted omega fitting range

* black

* black

* updated fitting ranges and data2features

* metaset validation fix

* updated impropers for PRO non PRO

* typo

* dihedral constant term revamp

* Add adjusted R2 fit parameter

* updated prior fitting for linear regression

* rename class

* metric choice, free parameters fix, and docs

* stats fix

* adjusted aic test and cleaned up Dihedral

* cleanup

* constant fix

* change order in saving model

* cath dataset path update

* pro specific prior adjusts and accomodation for new Dihedral

* constraint/script-level fixes

* default path updates

* t

* fix for prior specialization

* fix rigto compatibility with schnet

* fix rigto shape

* villin and wwdomain additions

* merge res

* notebbok clear

* pre merge wip

* prior specialization and initial pos spread fixes

* attach config fix

* vv

* add profiler option in base simulation

* set precission of the simulation

* PT sim dtype fixes

* DESRES wwdomain and non-force dataset handling

* hot fix so that h5_data works in the notebook
(no ddp)

* add types dependant out mlp

* DESRES wwdomain from bgmol fix

* fix bias in filter_network

* addition of possible internal lower cutoff in ExpNorm

* super fix

* expnormal revert

* added ExtendedExpNormalBasis

* init adjust

* last bias filter net revert

* fix pt script

* removed nvt cli jit depth command

* specific prior nvt script fix

* added gen_mol_configurations.py

* parser docs adjust

* added periodic prior yaml file

* Correction for repeated omega angle constraint

* merged with fix

* make sure index_mapping is a tensor

* add check

* fix shape for omega mapping index

* add nl generation, trajectory cg and
processing of the dimer dataset tools in examples/multi_datasets

* fix MLP compatibility with Schnet

* fixed langevin cli to allow single element beta list options

* legacy nls for gen_mol_configs

* removed CGNet dep. black.

* opep mapping ordering warning

* added CG pdb reordering tool

* variable/function name fixes

* --out default fix

* removed unneeded pruning

* update prior model in process_dimers

* make_opep_nls psi group fix

* mapping, constraint projection updates

* atom name adjustments

* lazy wip stash

* switched to full res specification for repulsion

* aggforce inclusion

* typo

* add support for mace activation function

* delta force production batch size and constraint finder data size reductions

* added optimized force mapping

* OverdampedSimulation attachment recursion error and test fixes

* remove unused functions

* removed unused imports

* debye nls wip

* fixed debye nls key

* dimers WIP

* generalization to multiple molecule pdbs

* temp commit

* added agg loaders

* gitignore

* lint

* small updates

* mut removes

* Create dihedrals_no_pbonds_force_aggregated_cath_opep_omega_fix_periodic_no_double_t erm_om_percentile_0.1_correct_hybrid_repul_only_ca_o.yml

add config

* Create dihedrals_no_pbonds_force_aggregated_cath_opep_omega_fix_periodic_no_double_term_om_percentile_0.1_correct_hybrid_repul_only_ca_o.yml

add configs

* Update dihedrals_no_pbonds_force_aggregated_cath_opep_omega_fix_periodic_no_double_t erm_om_percentile_0.1_correct_hybrid_repul_only_ca_o.yml

* Update dihedrals_no_pbonds_force_aggregated_cath_opep_omega_fix_periodic_no_double_t erm_om_percentile_0.1_correct_hybrid_repul_only_ca_o.yml

* Delete dihedrals_no_pbonds_force_aggregated_cath_opep_omega_fix_periodic_no_double_t erm_om_percentile_0.1_correct_hybrid_repul_only_ca_o.yml

* Create dihedrals_no_pbonds_force_aggregated_dimers_fix_periodic_no_double_term_om_percentile_0.1_correct_hybrid_repul_only_ca_o.yml

* Minor correction for PTSimulation

* Delete examples/multi_dataset/dihedrals_more_data_no_pbonds_cath_opep.yml

* Delete examples/multi_dataset/dihedrals_no_pbonds_cath_opep.yml

* Delete examples/multi_dataset/dihedrals_no_pbonds_cath_opep_omega_fix_periodic.yml

* Delete examples/multi_dataset/dihedrals_no_pbonds_only_opep.yml

* Delete examples/multi_dataset/gen_mol_configurations_legacy.py

* Delete examples/multi_dataset/generate_nl.py

* Delete examples/multi_dataset/process_dimers.py

* Delete examples/multi_dataset/sim_opts_no_scheduler_epoch_5.json

* Delete examples/multi_dataset/simulate_mlcg_trpcage_dihedrals.py

* Delete examples/multi_dataset/simulate_mlcg_dihedrals.py

* Delete examples/multi_dataset/simulate_mlcg_cln_prod.py

* Delete examples/multi_dataset/no_pbonds_cath_opep.yml

* Delete examples/multi_dataset/simulate_mlcg.py

* Delete examples/multi_dataset/sim_opts_cln_filter_fix_no_pbonds_epoch_19.json

* Delete examples/multi_dataset/data_gen_opts.yml

* Delete examples/multi_dataset/dihedrals_more_data_no_pbonds_cath_opep.yml

* Delete examples/multi_dataset/gen_mol_configurations_legacy.py

* Delete examples/multi_dataset/generate_nl.py

* Delete examples/multi_dataset/process_dimers.py

* Delete examples/multi_dataset/no_pbonds_cath_opep.yml

* Delete examples/multi_dataset/dihedrals_no_pbonds_cath_opep_omega_fix_periodic.yml

* Delete examples/multi_dataset/dihedrals_no_pbonds_cath_opep.yml

* Delete examples/multi_dataset/sim_opts_cln_filter_fix_no_pbonds_epoch_19.json

* Delete examples/multi_dataset/sim_opts_no_scheduler_epoch_5.json

* Delete examples/multi_dataset/simulate_mlcg_dihedrals.py

* Delete examples/multi_dataset/simulate_mlcg.py

* Delete examples/multi_dataset/simulate_mlcg_trpcage_dihedrals.py

* Delete examples/multi_dataset/dihedrals_no_pbonds_only_opep.yml

* Delete examples/multi_dataset/opeps_dataset.py

* Delete examples/multi_dataset/simulate_mlcg_cln_prod.py

* Create README.md for diverse datasets

* spelling fix for non-mut defaults in SumOut

* spelling fix for non-mut defaults in SumOut

* lint

---------

+++ mlcg/nn/cutoff.py
@@ -127,15 +127,21 @@ class CosineCutoff(_Cutoff):
                 + 1.0
             )
             # remove contributions below the cutoff radius
-            cutoffs = cutoffs * (distances < self.cutoff_upper).float()
-            cutoffs = cutoffs * (distances > self.cutoff_lower).float()
+            cutoffs = cutoffs * (distances < self.cutoff_upper).to(
+                distances.dtype
+            )
+            cutoffs = cutoffs * (distances > self.cutoff_lower).to(
+                distances.dtype
+            )
             return cutoffs
         else:
             cutoffs = 0.5 * (
                 torch.cos(distances * math.pi / self.cutoff_upper) + 1.0
             )
             # remove contributions beyond the cutoff radius
-            cutoffs = cutoffs * (distances < self.cutoff_upper).float()
+            cutoffs = cutoffs * (distances < self.cutoff_upper).to(
+                distances.dtype
+            )
             return cutoffs
